requirements.txt
aiogram==3.13.1
asyncpg==0.29.0
aiohttp==3.9.1
python-dotenv==1.0.0



